refactor: rename crates to cell-sheet-core and cell-sheet-tui
Rename for crates.io availability and consistent naming.
Set up trusted publishing workflow and add required package metadata.

Made-with: Cursor
