+++ .github/workflows/release.yml
@@ -6,6 +6,7 @@ on:
 
 permissions:
   contents: write
+  id-token: write
 
 env:
   CARGO_TERM_COLOR: always
@@ -54,7 +55,7 @@ jobs:
           key: ${{ matrix.target }}
 
       - name: Build
-        run: cargo build --release --target ${{ matrix.target }} -p cell-tui
+        run: cargo build --release --target ${{ matrix.target }} -p cell-sheet-tui
 
       - name: Package (unix)
         if: matrix.archive == 'tar.gz'
@@ -90,3 +91,28 @@ jobs:
         with:
           generate_release_notes: true
           files: artifacts/*
+
+  publish:
+    name: Publish to crates.io
+    needs: release
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: dtolnay/rust-toolchain@stable
+
+      - uses: rust-lang/crates-io-auth-action@v1
+        id: auth
+
+      - name: Publish cell-sheet-core
+        run: cargo publish -p cell-sheet-core
+        env:
+          CARGO_REGISTRY_TOKEN: ${{ steps.auth.outputs.token }}
+
+      - name: Wait for crates.io index
+        run: sleep 30
+
+      - name: Publish cell-sheet-tui
+        run: cargo publish -p cell-sheet-tui
+        env:
+          CARGO_REGISTRY_TOKEN: ${{ steps.auth.outputs.token }}
