+++ Cargo.toml
@@ -1,8 +1,11 @@
 [workspace]
-members = ["crates/cell-core", "crates/cell-tui"]
+members = ["crates/cell-sheet-core", "crates/cell-sheet-tui"]
 resolver = "2"
 
 [workspace.package]
 version = "0.1.0"
 edition = "2021"
 license = "MIT"
+repository = "https://github.com/garritfra/cell"
+homepage = "https://github.com/garritfra/cell"
+authors = ["Garrit Franke"]
