+++ crates/cell-sheet-core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cell-sheet-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+description = "Core spreadsheet engine with formulas, dependency tracking, and file I/O"
+
+[dependencies]
+csv = "1.3"
+
+[dev-dependencies]
+pretty_assertions = "1"
