+++ crates/cell-sheet-core/src/model.rs
