+++ crates/cell-sheet-tui/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "cell-sheet-tui"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+description = "A terminal spreadsheet editor with Vim-like keybindings"
+
+[[bin]]
+name = "cell"
+path = "src/main.rs"
+
+[dependencies]
+cell-sheet-core = { path = "../cell-sheet-core", version = "0.1.0" }
+ratatui = "0.30"
+crossterm = "0.29"
+clap = { version = "4", features = ["derive"] }
