+++ crates/cell-sheet-tui/src/viewport.rs
@@ -1,4 +1,4 @@
-use cell_core::model::CellPos;
+use cell_sheet_core::model::CellPos;
 
 pub struct Viewport {
     pub row_offset: usize,
