--- crates/cell-tui/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "cell-tui"
-version.workspace = true
-edition.workspace = true
-
-[[bin]]
-name = "cell"
-path = "src/main.rs"
-
-[dependencies]
-cell-core = { path = "../cell-core" }
-ratatui = "0.30"
-crossterm = "0.29"
-clap = { version = "4", features = ["derive"] }
